moving the generic view tests into their own file
